+++ addons/external.go
@@ -66,4 +66,11 @@ func init() {
 		},
 		Module: "github.com/invopop/gobl.sa.zatca",
 	})
+	tax.RegisterApprovedAddon(&tax.ExternalAddon{
+		Key: "pt-saft-v1",
+		Name: i18n.String{
+			i18n.EN: "Portugal SAF-T",
+		},
+		Module: "github.com/invopop/gobl.pt.saft",
+	})
 }